+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * (bc) Remove `ComponentsWithTags`, as it is replaced with `ComponentFilter`.
 * (bc) Remove ability to filter by component groups.
 * (improvement) Add internal `LocaleHelper` to unify locale detection.
+* (feature) Add `StoryMetaData::getTranslatedDocumentsMapping()` to fetch the required data to map hreflangs.
 
 
 3.0.0-beta.5